stlye mistakes corrected
