+++ src/main/java/corebasesyntax/model/Cat.java
@@ -1,0 +1,19 @@
+package corebasesyntax.model;
+
+public class Cat {
+    private int age;
+    private String name;
+    private String ownerName;
+
+    @Override
+    public String toString() {
+        return "Cat{"
+                + "age="
+                + age
+                + ", name='"
+                + name + '\''
+                + ", ownerName='"
+                + ownerName + '\''
+                + '}';
+    }
+}
