+++ src/main/java/corebasesyntax/model/Dog.java
@@ -1,0 +1,4 @@
+package corebasesyntax.model;
+
+public class Dog {
+}
